Use plugin event in integratin test
Probot 0.9.0 has switched to using Express for its server, which breaks
the `listening` event used by the integration test. Rather than relying
on this implementation detail, the Plugin now emits a `loaded` event
when it is loaded, and this event is used in the integration test to
indicate when the probot instance is ready to receive requests.